+++ src/ngMock/angular-mocks.js
@@ -1572,7 +1572,7 @@ function MockHttpExpectation(method, url, data, headers) {
     if (angular.isUndefined(data)) return true;
     if (data && angular.isFunction(data.test)) return data.test(d);
     if (data && angular.isFunction(data)) return data(d);
-    if (data && !angular.isString(data)) return angular.toJson(data) == d;
+    if (data && !angular.isString(data)) return angular.equals(data, angular.fromJson(d));
     return data == d;
   };
 
@@ -1655,21 +1655,6 @@ angular.mock.$TimeoutDecorator = function($delegate, $browser) {
    */
   $delegate.flush = function(delay) {
     $browser.defer.flush(delay);
-  };
-
-  /**
-   * @ngdoc method
-   * @name ngMock.$timeout#flushNext
-   * @methodOf ngMock.$timeout
-   * @description
-   *
-   * Flushes the next timeout in the queue and compares it to the provided delay
-   *
-   * @param {number=} expectedDelay the delay value that will be asserted against the delay of the
-   *                  next timeout function
-   */
-  $delegate.flushNext = function(expectedDelay) {
-    $browser.defer.flushNext(expectedDelay);
   };
 
   /**