+++ test/ngRoute/directive/ngViewSpec.js
@@ -514,6 +514,76 @@ describe('ngView', function() {
   });
 });
 
+describe('ngView and transcludes', function() {
+  var element, directive;
+
+  beforeEach(module('ngRoute', function($compileProvider) {
+    element = null;
+    directive = $compileProvider.directive;
+  }));
+
+  afterEach(function() {
+    if (element) {
+      dealoc(element);
+    }
+  });
+
+  it('should allow access to directive controller from children when used in a replace template', function() {
+    var controller;
+    module(function($routeProvider) {
+      $routeProvider.when('/view', {templateUrl: 'view.html'});
+      directive('template', function() {
+        return {
+          template: '<div ng-view></div>',
+          replace: true,
+          controller: function() {
+            this.flag = true;
+          }
+        };
+      });
+
+      directive('test', function() {
+        return {
+          require: '^template',
+          link: function(scope, el, attr, ctrl) {
+            controller = ctrl;
+          }
+        };
+      });
+    });
+    inject(function($compile, $rootScope, $httpBackend, $location) {
+      $httpBackend.expectGET('view.html').respond('<div><div test></div></div>');
+      element = $compile('<div><div template></div></div>')($rootScope);
+      $location.url('/view');
+      $rootScope.$apply();
+      $httpBackend.flush();
+      expect(controller.flag).toBe(true);
+    });
+  });
+
+  it("should compile it's content correctly (although we remove it later)", function() {
+    var testElement;
+    module(function($compileProvider, $routeProvider) {
+      $routeProvider.when('/view', {template: ' '});
+      var directive = $compileProvider.directive;
+      directive('test', function() {
+        return {
+          link: function(scope, element) {
+            testElement = element;
+          }
+        };
+      });
+    });
+    inject(function($compile, $rootScope, $location) {
+      element = $compile('<div><div ng-view><div test someAttr></div></div></div>')($rootScope);
+      $location.url('/view');
+      $rootScope.$apply();
+      expect(testElement[0].nodeName).toBe('DIV');
+    });
+
+  });
+});
+
 describe('ngView animations', function() {
   var body, element, $rootElement;
 
@@ -619,7 +689,6 @@ describe('ngView animations', function() {
         item = $animate.flushNext('enter').element;
 
         $animate.flushNext('addClass').element;
-        $animate.flushNext('addClass').element;
 
         expect(item.hasClass('classy')).toBe(true);
 
@@ -638,7 +707,6 @@ describe('ngView animations', function() {
         $animate.flushNext('enter').element;
         item = $animate.flushNext('leave').element;
 
-        $animate.flushNext('addClass').element;
         $animate.flushNext('addClass').element;
 
         expect(item.hasClass('boring')).toBe(true);